Pass fastMsgIdFn to GossipSub constructor
