+++ package.json
@@ -41,7 +41,6 @@
   "homepage": "https://github.com/ChainSafe/js-libp2p-gossipsub#readme",
   "dependencies": {
     "@types/debug": "^4.1.7",
-    "@chainsafe/as-sha256": "^0.2.4",
     "debug": "^4.3.1",
     "denque": "^1.5.0",
     "err-code": "^3.0.1",
@@ -53,6 +52,7 @@
   },
   "devDependencies": {
     "@chainsafe/libp2p-noise": "^4.1.1",
+    "@chainsafe/as-sha256": "^0.2.4",
     "@types/chai": "^4.2.3",
     "@types/mocha": "^8.2.2",
     "@typescript-eslint/eslint-plugin": "^3.0.2",
