+++ test/utils/index.js
@@ -1,11 +1,9 @@
 'use strict'
 
 const { expect } = require('chai')
-const { messageIdToString } = require('../../src/utils/messageIdToString')
 const FloodSub = require('libp2p-floodsub')
 const PeerId = require('peer-id')
 const delay = require('delay')
-const { fromString: uint8ArrayFromString } = require('uint8arrays/from-string')
 
 exports.first = (map) => map.values().next().value
 
@@ -41,23 +39,11 @@ exports.createFloodsubNode = createFloodsubNode
 for (const [k, v] of Object.entries({
   ...require('./create-peer'),
   ...require('./create-gossipsub'),
-  ...require('./make-test-message')
+  ...require('./make-test-message'),
+  ...require('./msgId'),
 })) {
   exports[k] = v
 }
-
-const getMsgId = (msg) => {
-  const from = uint8ArrayFromString(msg.from)
-  const seqno = uint8ArrayFromString(msg.seqno)
-  const result = new Uint8Array(from.length + seqno.length)
-  result.set(from, 0)
-  result.set(seqno, from.length)
-  return result
-}
-
-exports.getMsgId = getMsgId
-
-exports.getMsgIdStr = (msg) => messageIdToString(getMsgId(msg))
 
 exports.waitForAllNodesToBePeered = async (peers, attempts = 10, delayMs = 100) => {
   const nodeIds = peers.map(peer => peer.peerId.toB58String())
